fix(browser): align cookie expires parsing and recover cookies on startup
Follow-up to 796fd6d. Re-review found two remaining gaps in the session
cookie lifecycle; both fixed and verified end-to-end.

1) create/update API request structs still used int64 for expires, so a
   fractional value (e.g. a pasted browser export) failed ShouldBindJSON
   with 400. Now float64 + int64() truncation, matching the import/recovery
   path fixed in 796fd6d.

2) Cookies were only recovered on graceful SIGTERM (Bot.Close). A hard kill
   (SIGKILL/crash/OOM) left the wrapper's period-saveState file orphaned and
   never imported to DB, so they were lost on restart. Added a startup
   recovery step in setupBrowserMCP that merges the container's residual
   state file into DB (upsert by bot_id+domain+name+path, DB-only rows
   preserved) before the loader overwrites the file. Idempotent for graceful
   restarts (file == DB); closes the non-graceful-shutdown gap.

Verified on bot-2d8f9b087270da0bcfe177a5:
- startup recovery log 'browser cookies recovered at startup'; a cookie placed
  only in the container state file (with a fractional expires 1789999999.5)
  appeared in DB truncated to 1789999999.
- create + update APIs accept a fractional expires and store the int64 floor.

+++ api/handler_bot_browser.go
@@ -65,14 +65,14 @@ func (s *Server) handleGetBotBrowserCookie(c *gin.Context) {
 func (s *Server) handleCreateBotBrowserCookie(c *gin.Context) {
 	botID := c.Param("id")
 	var req struct {
-		Domain   string `json:"domain"`
-		Name     string `json:"name"`
-		Value    string `json:"value"`
-		Path     string `json:"path"`
-		Expires  int64  `json:"expires"`
-		HTTPOnly bool   `json:"httpOnly"`
-		Secure   bool   `json:"secure"`
-		SameSite string `json:"sameSite"`
+		Domain   string  `json:"domain"`
+		Name     string  `json:"name"`
+		Value    string  `json:"value"`
+		Path     string  `json:"path"`
+		Expires  float64 `json:"expires"` // 与 import/recovery 一致，接受浮点秒（浏览器导出常带小数）
+		HTTPOnly bool    `json:"httpOnly"`
+		Secure   bool    `json:"secure"`
+		SameSite string  `json:"sameSite"`
 	}
 	if err := c.ShouldBindJSON(&req); err != nil {
 		Fail(c, errs.BadRequest("invalid request body: "+err.Error()))
@@ -92,7 +92,7 @@ func (s *Server) handleCreateBotBrowserCookie(c *gin.Context) {
 		Name:     req.Name,
 		Value:    req.Value,
 		Path:     req.Path,
-		Expires:  req.Expires,
+		Expires:  int64(req.Expires),
 		HTTPOnly: req.HTTPOnly,
 		Secure:   req.Secure,
 		SameSite: req.SameSite,
@@ -119,14 +119,14 @@ func (s *Server) handleUpdateBotBrowserCookie(c *gin.Context) {
 		return
 	}
 	var req struct {
-		Domain   *string `json:"domain"`
-		Name     *string `json:"name"`
-		Value    *string `json:"value"`
-		Path     *string `json:"path"`
-		Expires  *int64  `json:"expires"`
-		HTTPOnly *bool   `json:"httpOnly"`
-		Secure   *bool   `json:"secure"`
-		SameSite *string `json:"sameSite"`
+		Domain   *string  `json:"domain"`
+		Name     *string  `json:"name"`
+		Value    *string  `json:"value"`
+		Path     *string  `json:"path"`
+		Expires  *float64 `json:"expires"` // 与 import/recovery 一致，接受浮点秒
+		HTTPOnly *bool    `json:"httpOnly"`
+		Secure   *bool    `json:"secure"`
+		SameSite *string  `json:"sameSite"`
 	}
 	if err := c.ShouldBindJSON(&req); err != nil {
 		Fail(c, errs.BadRequest("invalid request body: "+err.Error()))
@@ -145,7 +145,7 @@ func (s *Server) handleUpdateBotBrowserCookie(c *gin.Context) {
 		ck.Path = *req.Path
 	}
 	if req.Expires != nil {
-		ck.Expires = *req.Expires
+		ck.Expires = int64(*req.Expires)
 	}
 	if req.HTTPOnly != nil {
 		ck.HTTPOnly = *req.HTTPOnly
